Chore: Refactor the environment state from the state sync
